refactor: move use-pdf-page-renderer to shared hooks directory
